feat(emus): add method to include rom launcher in all emu launch scripts
fix(emus): #121/remove launch items with no launch script
refactor(emu) split cpu profile management in to separate class
